backend/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
slowapi==0.1.9
pytest==7.4.3
httpx==0.25.2
reportlab==4.0.7
pillow==10.1.0
